Move ActionButtons component back to "new" as it's specific to the new component
